moved files to 2-Webpack and added copy text plugin for css files
