+++ 2-Webpack/src/styles.css
@@ -1,0 +1,148 @@
+button{
+    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
+    border: none;
+    background-color: white;
+    border-radius: 10px;
+    height:40px;
+    margin: 5px  
+}
+button:hover{
+    transform: scale(1.3)
+}
+input{
+    margin: 0px 10px;
+}
+input::-webkit-input-placeholder {
+    color: #b2cde0
+ }
+input:hover{
+    transform: scale(1.1);
+}
+.page-body{
+    display:flex;
+    flex-direction:column;
+    height: 100vh;
+}
+.header{
+    background-color: lightblue;
+    padding: 20px 20px 10px 20px;
+    border-bottom: solid 3px rgb(0, 140, 255);
+}
+.main-body{
+    display:flex;
+    flex-direction:row;
+    height: 100vh;
+}
+.menu{
+    background-color: whitesmoke; 
+    padding: 10px;
+    border-right: solid 2px grey;
+}
+.menu-buttons{
+    display: flex;
+    justify-content: space-around;
+}
+.menu-item{
+    display: flex;
+    flex-direction: row;
+    justify-content: end;
+    margin: 5px;
+    width: 100%;
+}
+.content-input{
+    height:50px;
+}
+.item{
+    display:flex; 
+    flex-direction:column; 
+    align-items: end;
+}
+.display{
+    min-width: 40%;
+    max-width: 60%;
+    display: flex;
+    flex-direction: column;
+    margin: 10px;
+    justify-content: start;
+    align-items: center;
+}
+.todo-item{
+    display: flex; 
+    flex-direction: row;
+    justify-content: space-between;
+    padding:5px; 
+    border-radius: 10px; 
+    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
+    margin-bottom: 12px;
+}
+
+.item-checkbox{
+    margin-right: 10px;
+}
+.item-checkbox:checked ~div.todo-item-text > .todo-item-title {
+    text-decoration: line-through;
+}
+.item-checkbox:checked ~div.todo-item-text > .todo-item-subtitle {
+    text-decoration: line-through;
+}
+.item-checkbox:checked ~div.todo-item-text > .todo-item-content {
+    display: none;
+}
+.todo-item-text{
+    display: flex; 
+    flex-direction: column;
+    align-self: stretch;
+}
+.todo-item-buttons{
+    display: flex; 
+    flex-direction: row; 
+    justify-content: space-between; 
+    align-items: center;
+}
+.item-edit-button{
+    font-family: 'Material Icons';
+    font-size: 30px;
+    text-align: center;
+    color: white;
+    width:40px;
+    height:40px;
+    border-radius: 30%;
+    background-color: greenyellow;
+}
+.item-edit-button:hover{
+    background-color: seagreen;  
+}
+.item-delete-button{
+    font-family: 'Material Icons';
+    font-size: 30px;
+    text-align: center;
+    color: white;
+    width:40px;
+    height:40px;
+    border-radius: 30%;
+    background-color: tomato;
+}
+.item-delete-button:hover{
+    background-color: crimson;
+}
+.item-checkbox{
+    align-self: center;
+}
+.item-checkbox:hover{
+    transform: scale(1.3);
+}
+.todo-item-title{
+    font-size: 20px;
+    font-weight: bold;
+}
+.todo-item-subtitle{
+    font-size: 20px;
+    color: grey;
+    font-weight: thin;
+    padding-left: 15px;
+}
+.todo-item-content{
+    font-weight: thin;
+    padding-left: 15px;
+    color: grey;
+}
